+++ common/cpp/src/database/mongodb_client.h
@@ -33,6 +33,12 @@ class BsonDestroyFunctor {
 // using _bson_t here to avoid GNU compiler warnings
 using bson_p = std::unique_ptr<_bson_t, BsonDestroyFunctor>;
 
+enum class GetRecordMode {
+  kById,
+  kLast,
+  kEarliest
+};
+
 class MongoDBClient final : public Database {
   public:
     MongoDBClient();
@@ -44,6 +50,7 @@ class MongoDBClient final : public Database {
     Error GetById(const std::string& collection, uint64_t id, FileInfo* file) const override;
     Error GetDataSetById(const std::string& collection, uint64_t set_id, uint64_t id, FileInfo* file) const override;
     Error GetStreamInfo(const std::string& collection, StreamInfo* info) const override;
+    Error GetLastStream(StreamInfo* info) const override;
     ~MongoDBClient() override;
   private:
     mongoc_client_t* client_{nullptr};
@@ -61,7 +68,9 @@ class MongoDBClient final : public Database {
     Error InsertBsonDocument(const bson_p& document, bool ignore_duplicates) const;
     Error UpdateBsonDocument(uint64_t id, const bson_p& document, bool upsert) const;
     Error AddBsonDocumentToArray(bson_t* query, bson_t* update, bool ignore_duplicates) const;
-    Error GetRecordFromDb(const std::string& collection, uint64_t id, bool ignore_id_return_last, std::string* res) const;
+    Error GetRecordFromDb(const std::string& collection, uint64_t id, GetRecordMode mode, std::string* res) const;
+    Error UpdateStreamInfo(const char *str,StreamInfo* info) const;
+
 };
 
 }