+++ consumer/api/cpp/include/consumer/data_broker.h
@@ -4,6 +4,7 @@
 #include <memory>
 #include <string>
 #include <vector>
+#include <chrono>
 
 #include "common/data_structs.h"
 #include "common/error.h"
@@ -73,8 +74,8 @@ class DataBroker {
      */
     virtual NetworkConnectionType CurrentConnectionType() const = 0;
 
-    //! Set list of substreams
-    virtual std::vector<std::string> GetSubstreamList(Error* err) = 0;
+    //! Get list of substreams, set from to "" to get all substreams
+    virtual StreamInfos GetSubstreamList(std::string from, Error* err) = 0;
 
     //! Get current number of datasets
     /*!