+++ receiver/src/request_handler/request_factory.h
@@ -5,6 +5,7 @@
 #include "../file_processors/write_file_processor.h"
 #include "../file_processors/receive_file_processor.h"
 #include "request_handler_db_stream_info.h"
+#include "request_handler_db_last_stream.h"
 
 namespace asapo {
 
@@ -24,6 +25,7 @@ class RequestFactory {
     RequestHandlerReceiveMetaData request_handler_receive_metadata_;
     RequestHandlerDbWrite request_handler_dbwrite_{kDBDataCollectionNamePrefix};
     RequestHandlerDbStreamInfo request_handler_db_stream_info_{kDBDataCollectionNamePrefix};
+    RequestHandlerDbLastStream request_handler_db_last_stream_{kDBDataCollectionNamePrefix};
     RequestHandlerDbMetaWrite request_handler_db_meta_write_{kDBMetaCollectionName};
     RequestHandlerAuthorize request_handler_authorize_;
     RequestHandlerDbCheckRequest request_handler_db_check_{kDBDataCollectionNamePrefix};;