+++ tests/automatic/consumer/consumer_api/consumer_api.cpp
@@ -134,13 +134,19 @@ void TestSingle(const std::unique_ptr<asapo::DataBroker>& broker, const std::str
     M_AssertTrue(err == nullptr, "GetNext stream2 no error");
     M_AssertTrue(fi.name == "21", "GetNext stream2 filename");
 
-    auto substreams = broker->GetSubstreamList(&err);
+    auto substreams = broker->GetSubstreamList("",&err);
     M_AssertTrue(err == nullptr, "GetSubstreamList no error");
     M_AssertTrue(substreams.size() == 3, "substreams.size");
-    M_AssertTrue(substreams[0] == "default", "substreams.name1");
-    M_AssertTrue(substreams[1] == "stream1", "substreams.name2");
-    M_AssertTrue(substreams[2] == "stream2", "substreams.name3");
-
+    M_AssertTrue(substreams[0].name == "default", "substreams0.name1");
+    M_AssertTrue(substreams[1].name == "stream1", "substreams1.name2");
+    M_AssertTrue(substreams[2].name == "stream2", "substreams2.name3");
+    std::cout<<substreams[0].Json(false)<<std::endl;
+    std::cout<<substreams[1].Json(false)<<std::endl;
+    std::cout<<substreams[2].Json(false)<<std::endl;
+    M_AssertTrue(asapo::NanosecsEpochFromTimePoint(substreams[0].timestamp_created) == 0, "substreams0.timestamp");
+    M_AssertTrue(asapo::NanosecsEpochFromTimePoint(substreams[0].timestamp_lastentry) == 0, "substreams0.timestamp lastentry not set");
+    M_AssertTrue(asapo::NanosecsEpochFromTimePoint(substreams[1].timestamp_created) == 1000, "substreams1.timestamp");
+    M_AssertTrue(asapo::NanosecsEpochFromTimePoint(substreams[2].timestamp_created) == 2000, "substreams2.timestamp");
 // acknowledges
 
     auto id = broker->GetLastAcknowledgedTulpeId(group_id, &err);
@@ -254,8 +260,6 @@ void TestAll(const Args& args) {
     }
 
 }
-
-
 
 int main(int argc, char* argv[]) {